Finish set list page functionality, adjust API to disallow for invalid registrations

+++ capstone-computing-project/src/auth/setlist.js
@@ -27,15 +27,12 @@ module.exports.getSetList = async (req, res) => {
     const { startDate, endDate } = req.query;
 
     // Check if the dates are correctly passed through
+    console.log("Fetching set list reservations...");
     console.log('Start Date:', startDate, 'End Date:', endDate);
 
     authenticateJWT(req, res, () => {
-        // console.log("Validated successfully");
-        // const { token } = req.body;
-        // console.log("User info");
-        // console.log(req.user);
+        // Get current user's CWID
         const cwid = req.user.id;
-        console.log(cwid);
 
         // Dynamically pass the startDate and endDate from the request instead of hardcoded values
         db.query('SELECT SetList.Date, User.Fname, User.Lname, User.Email, User.CWID FROM User INNER JOIN SetList ON User.CWID = SetList.CWID WHERE SetList.Date >= ? AND SetList.Date < ?', 
@@ -45,11 +42,8 @@ module.exports.getSetList = async (req, res) => {
                     return res.status(500).json({ message: 'Database error' });
                 }
         
-                // Log the results to verify how many rows are returned
-                console.log('Results from DB:', results);
                 // For each item in results, determine if it was registered by the given user
                 for (var i = 0; i < results.length; i++) {
-                    console.log(results[i].CWID);
                     if (results[i].CWID == cwid) {
                         results[i].RegisteredBy = "you";
                     } else {
@@ -57,7 +51,7 @@ module.exports.getSetList = async (req, res) => {
                     }
                     delete results[i].CWID;
                 }
-                console.log('After delete:', results);
+                
                 res.status(200).json(results);
         });
     });
@@ -92,28 +86,27 @@ module.exports.getSetList = async (req, res) => {
 // ];
 
 module.exports.registerReservation = async (req, res) => {
-    console.log("Attempt to reserve!");
+    console.log("Reservation started");
 
-    // In here, we need to get CWID for user. Get this from token
-    console.log(req.body);
     const { reserveDate } = req.body;
+
     // Authenticate user is logged in before allowing reservation
-    
-    // Getting back 403 on this. Issue with my token?
     authenticateJWT(req, res, () => {
-        // console.log("Validated successfully");
-        // const { token } = req.body;
-        // console.log("User info");
-        // console.log(req.user);
+
         const cwid = req.user.id;
-        console.log(cwid);
-        console.log(reserveDate);
         const date = new Date(reserveDate);
-        console.log(date.toString());
+
+        // Disallow reservation if reservation date is before current time
+        if (date < new Date()) {
+            return res.status(403).json({ error: 'Failed to create reservation', message: "Cannot make a reservation before the current time!" });
+        }
+
+        // Format date string to what database is expecting
         const dateString = date.getFullYear() + "-" + (date.getMonth() + 1).toString() + "-" + date.getDate() + " " + date.getHours() + ":" + date.getMinutes() + ":00";
-        // return error if date already exists
+        
         const query = 'INSERT INTO SetList (Date, CWID) VALUES (?, ?)';
         db.query(query, [dateString, cwid], (err, result) => {
+            // return error if date already exists
             if (err) {
                 return res.status(500).json({ error: 'Failed to make reservation' });
             }
@@ -127,24 +120,25 @@ module.exports.registerReservation = async (req, res) => {
 
 module.exports.deleteReservation = async (req, res) => {
     // Authenticate user is logged in before allowing reservation
-    console.log("INSIDE deleteReservation");
+    console.log("Reservation cancellation");
     const { reserveDate } = req.body;
 
-    // Authenticate user is logged in before allowing reservation
-    // console.log(req.body);
+    // Authenticate user is logged in before allowing delete of reservation
     authenticateJWT(req, res, () => {
-        // console.log("Validated successfully");
-        // const { token } = req.body;
-        // console.log("User info");
-        // console.log(req.user);
 
         // NOTE: Should add check in API here to disallow registrations before current time.
 
         const cwid = req.user.id;
-        console.log(cwid);
-        console.log(reserveDate);
+        // console.log(cwid);
+        // console.log(reserveDate);
         const date = new Date(reserveDate);
-        console.log(date.toString());
+
+        // Disallow deletion if reservation date is before current time
+        if (date < new Date()) {
+            return res.status(403).json({ error: 'Failed to delete reservation', message: "Cannot delete a reservation before the current time!" });
+        }
+
+        // console.log(date.toString());
         const dateString = date.getFullYear() + "-" + (date.getMonth() + 1).toString() + "-" + date.getDate() + " " + date.getHours() + ":" + date.getMinutes() + ":00";
         // return error if date already exists
         const query = 'DELETE FROM SetList WHERE (Date = ? AND CWID = ?)';
@@ -158,40 +152,4 @@ module.exports.deleteReservation = async (req, res) => {
     
     
     res.status(200);
-
-    // In here, we need to get CWID for user
-    console.log("Attempt to delete!");
-    res.status(200);
 }
-
-// exports.addMeetingNote = (req, res) => {
-//     const { title, content } = req.body;
-
-//     const currentDate = new Date();
-//     const formattedDate = currentDate.toISOString().split('T')[0];
-
-//     const query = 'INSERT INTO meeting_notes (title, content, date) VALUES (?, ?, ?)';
-//     db.query(query, [title, content, formattedDate], (err, result) => {
-//         if (err) {
-//             return res.status(500).json({ error: 'Failed to add note' });
-//         }
-//         res.status(201).json({
-//             id: result.insertId,
-//             title: title,
-//             content: content,
-//             date: formattedDate,
-//         });
-//     });
-// };
-
-// exports.deleteMeetingNote = (req, res) => {
-//     const { id } = req.params;
-//     const query = 'DELETE FROM meeting_notes WHERE id = ?';
-
-//     db.query(query, [id], (err, result) => {
-//         if (err) {
-//             return res.status(500).json({ error: 'Failed to delete note' });
-//         }
-//         res.json({ message: 'Note deleted successfully' });
-//     });
-// };
